ml-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
openai
python-dotenv
python-multipart
